requirements.txt
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
filelock==3.16.1
fsspec==2024.10.0
gdown==5.2.0
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.10.0.84
pillow==11.0.0
PySocks==1.7.1
requests==2.32.3
soupsieve==2.6
sympy==1.13.3
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
